% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{absurdo}
\alias{absurdo}
\title{Reduccion al absurdo}
\usage{
absurdo(AB, CA, valor)
}
\arguments{
\item{AB}{Corresponde a la primera premisa.}

\item{CA}{Corresponde a la segunda premisa.}

\item{valor}{Corresponde a la forma de la figura.}
}
\description{
Reduccion al absurdo
}
\examples{
reductio1 <- absurdo("AB", "CA", "AAA");
reductio2 <- absurdo("BA", "CA", "EAE");
}
