% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{Boecio}
\alias{Boecio}
\title{Cuadro de Boecio o de oposición de los juicios}
\usage{
Boecio(primer_termino, segundo_termino, tipo_proposicion)
}
\arguments{
\item{primer_termino}{Corresponde al sujeto de la proposición original entre comillas.}

\item{segundo_termino}{Corresponde al predicado de la proposición original entre comillas.}

\item{tipo_proposicion}{A, E, I, O, según la notación de las proposiciones.}
}
\value{
La proposición original, su opuesta, subordinada y contradictoria.
}
\description{
Cuadro de Boecio o de oposición de los juicios
}
\examples{
Boecio1 <- Boecio("Hombre", "Humano", "O");
Boecio2 <- Boecio("Caballo", "Animal", "Universal Afirmativa");
}
