% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{contradiccion}
\alias{contradiccion}
\title{Contradicción}
\usage{
contradiccion(primer_termino, segundo_termino, tipo_proposicion)
}
\arguments{
\item{primer_termino}{Corresponde al sujeto de la proposición original entre comillas.}

\item{segundo_termino}{Corresponde al predicado de la proposición original entre comillas.}

\item{tipo_proposicion}{A, E, I, O, según la notación de las proposiciones.}
}
\value{
La proposición original y su contradictoria.
}
\description{
Devuelve la proposición original y su contradicción para los términos y el tipo de proposición específicados.
}
\examples{
contradiccion1 <- contradiccion("Hombre", "Humano", "O");
contradiccion2 <- contradiccion("Caballo", "Animal", "Universal Afirmativa");
}
