% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{conversion}
\alias{conversion}
\title{Regla de conversión}
\usage{
conversion(primer_termino, segundo_termino, tipo_proposicion)
}
\arguments{
\item{primer_termino}{Corresponde al sujeto de la proposición original entre comillas.}

\item{segundo_termino}{Corresponde al predicado de la proposición original entre comillas.}

\item{tipo_proposicion}{A, E, I, O, según la notación de las proposiciones.}
}
\value{
La proposición original y su conversa.
}
\description{
Devuelve la proposición original y su conversa para los términos y el tipo de proposición específicados.
}
\examples{
conversion1 <- conversion("Hombre", "Humano", "O");
conversion2 <- conversion("Caballo", "Animal", "Universal Afirmativa");
}
