% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{figura_silogismo}
\alias{figura_silogismo}
\title{Figura Silogismo}
\usage{
figura_silogismo(mayor, medio, menor, modo, figura)
}
\arguments{
\item{mayor}{El término mayor entre comillas.}

\item{medio}{El término medio entre comillas.}

\item{menor}{El término menor entre comillas.}

\item{modo}{El modo del silogismo entre comillas.}

\item{figura}{La figura del silogismo entre comillas o en valor númerico.}
}
\value{
La figura del silogismo según el modo indicado, en caso de ser válido.
}
\description{
Si el modo es válido, devuelve la figura indicada del silogismo a partir de los términos definidos.
Si el modo es inválido, devuelve una advertencia.
}
\examples{
silogismo1 <- figura_silogismo("Hombre", "Animal", "Piedra", "CAMENES", "Cuarto");
silogismo2 <- figura_silogismo("Hombre", "Animal", "Piedra", "AAA", 1);
}
