% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{forma_axiomatica}
\alias{forma_axiomatica}
\title{Forma axiomática}
\usage{
forma_axiomatica(modo, figura)
}
\arguments{
\item{modo}{El modo del silogismo entre comillas.}

\item{figura}{La figura del silogismo entre comillas o en valor númerico.}
}
\value{
La figura axiomática del primer silogismo para el modo y figura especificados, en caso de ser válidos.
}
\description{
Si el modo es válido, devuelve la forma axiomática de la primera figura del silogismo.
}
\examples{
forma_primer_silogismo1 <- forma_axiomatica("BAROCO", "Segunda");
forma_primer_silogismo2 <- forma_axiomatica("IAI", "Cuarta");
}
