% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{forma_reduccion_absurdo}
\alias{forma_reduccion_absurdo}
\title{Forma axiomática para reducción al absurdo}
\usage{
forma_reduccion_absurdo(modo, figura)
}
\arguments{
\item{modo}{El modo del silogismo entre comillas.}

\item{figura}{La figura del silogismo entre comillas o en valor númerico.}
}
\value{
La forma axiomática del primer silogismo para resolver por reducción al absurdo.
}
\description{
Forma axiomática para reducción al absurdo
}
\examples{
forma_absurdo_silogismo1 <- forma_reduccion_absurdo("BAROCO", "Segunda");
forma_absurdo_silogismo2 <- forma_reduccion_absurdo("IAI", "Cuarta");
}
