% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{oposicion}
\alias{oposicion}
\title{Oposición}
\usage{
oposicion(primer_termino, segundo_termino, tipo_proposicion)
}
\arguments{
\item{primer_termino}{Corresponde al sujeto de la proposición original entre comillas.}

\item{segundo_termino}{Corresponde al predicado de la proposición original entre comillas.}

\item{tipo_proposicion}{A, E, I, O, según la notación de las proposiciones.}
}
\value{
La proposición original y su opuesta.
}
\description{
Devuelve la proposición original y su opuesta para los términos y el tipo de proposición específicados.
}
\examples{
oposicion1 <- oposicion("Hombre", "Humano", "O");
oposicion2 <- oposicion("Caballo", "Animal", "Universal Afirmativa");
}
