% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{primer_silogismo}
\alias{primer_silogismo}
\title{Primera figura del silogismo}
\usage{
primer_silogismo(mayor, medio, menor, modo)
}
\arguments{
\item{mayor}{El término mayor entre comillas.}

\item{medio}{El término medio entre comillas.}

\item{menor}{El término menor entre comillas.}
}
\value{
La primera figura del silogismo según el modo indicado, en caso de ser válido.
}
\description{
Primera figura del silogismo
}
\examples{
silogismo1 <- primer_silogismo("Hombre", "Animal", "Piedra", "BARBARA");
silogismo2 <- primer_silogismo("Hombre", "Animal", "Piedra", "EAE");
}
