% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{segundo_silogismo}
\alias{segundo_silogismo}
\title{Segunda figura del silogismo}
\usage{
segundo_silogismo(mayor, medio, menor, modo)
}
\arguments{
\item{mayor}{El término mayor entre comillas.}

\item{medio}{El término medio entre comillas.}

\item{menor}{El término menor entre comillas.}
}
\value{
La segunda figura del silogismo según el modo indicado, en caso de ser válido.
}
\description{
Segunda figura del silogismo
}
\examples{
silogismo1 <- segundo_silogismo("Hombre", "Animal", "Piedra", "CAMESTRES");
silogismo2 <- segundo_silogismo("Hombre", "Animal", "Piedra", "EIO");
}
