% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{tercer_silogismo}
\alias{tercer_silogismo}
\title{Tercera figura del silogismo}
\usage{
tercer_silogismo(mayor, medio, menor, modo)
}
\arguments{
\item{mayor}{El término mayor entre comillas.}

\item{medio}{El término medio entre comillas.}

\item{menor}{El término menor entre comillas.}
}
\value{
La tercera figura del silogismo según el modo indicado, en caso de ser válido.
}
\description{
Tercera figura del silogismo
}
\examples{
silogismo1 <- tercer_silogismo("Hombre", "Animal", "Piedra", "DATISIS");
silogismo2 <- tercer_silogismo("Hombre", "Animal", "Piedra", "EAO");
}
