% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{absurdo}
\alias{absurdo}
\title{Reduccion al absurdo}
\usage{
absurdo(AB, CA, valor)
}
\arguments{
\item{AB}{Corresponde a la primera premisa.}

\item{CA}{Corresponde a la segunda premisa.}

\item{valor}{Corresponde a la forma de la figura.}
}
\description{
A partir de la forma de las primera y segunda premisa, asi como la figura del
silogismo dice que forma debe tener el silogismo a evaluar en reduccion al absurdo,
o si este no es posible resolver por dicho metodo. Todavia no realiza la demostracion
completa de la reduccion al absurdo
}
\examples{
reductio1 <- absurdo("AB", "CA", "AAA");
reductio2 <- absurdo("BA", "CA", "EAE");
}
