% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silogismo.R
\name{cuarto_silogismo}
\alias{cuarto_silogismo}
\title{Cuarta figura del silogismo}
\usage{
cuarto_silogismo(mayor, medio, menor, modo)
}
\arguments{
\item{mayor}{El término mayor entre comillas.}

\item{medio}{El término medio entre comillas.}

\item{menor}{El término menor entre comillas.}
}
\value{
La cuarta figura del silogismo según el modo indicado, en caso de ser válido.
}
\description{
Cuarta figura del silogismo
}
\examples{
silogismo1 <- cuarto_silogismo("Hombre", "Animal", "Piedra", "CAMENES");
silogismo2 <- cuarto_silogismo("Hombre", "Animal", "Piedra", "EAO");
}
